# nbody.mojo - remove unnecessary memory writes and @unroll inner loop