--- a/examples/nbody.mojo
+++ b/examples/nbody.mojo
@@ -61,8 +61,8 @@ fn offset_momentum(inout bodies: StaticTuple[NUM_BODIES, Planet]):
 fn advance(inout bodies: StaticTuple[NUM_BODIES, Planet], dt: Float64):
     @unroll
     for i in range(NUM_BODIES):
+        var body_i = bodies[i]
         for j in range(NUM_BODIES - i - 1):
-            var body_i = bodies[i]
             var body_j = bodies[j + i + 1]
             let diff = body_i.pos - body_j.pos
             let diff_sqr = (diff * diff).reduce_add()
@@ -71,8 +71,9 @@ fn advance(inout bodies: StaticTuple[NUM_BODIES, Planet], dt: Float64):
             body_i.velocity -= diff * body_j.mass * mag
             body_j.velocity += diff * body_i.mass * mag
 
-            bodies[i] = body_i
             bodies[j + i + 1] = body_j
+
+        bodies[i] = body_i
 
     @unroll
     for i in range(NUM_BODIES):