--- a/examples/nbody.mojo
+++ b/examples/nbody.mojo
@@ -63,9 +63,7 @@ fn advance(inout bodies: StaticTuple[NUM_BODIES, Planet], dt: Float64):
     for i in range(NUM_BODIES):
         var body_i = bodies[i]
 
-        # @unroll(NUM_BODIES - 1) errors.
-        # Bug reported at https://github.com/modularml/mojo/pull/1534
-        @unroll(4)
+        @unroll(NUM_BODIES - 1)
         for j in range(NUM_BODIES - i - 1):
             var body_j = bodies[j + i + 1]
             let diff = body_i.pos - body_j.pos