Edit: I found a second optimisation unrolling the inner loop so the total improvement is 6.53 -> 1.13 seconds on M1 Pro.  That change is discussed in the next comment.

------------------
On M1 Pro the result of `benchmark()` falls from 6.53 to 4.40 seconds.  I realise this code is reproducing a specific benchmark but some of the other competitors do this optimisation so I think it is fine.

One of the [Julia code examples](https://benchmarksgame-team.pages.debian.net/benchmarksgame/program/nbody-julia-8.html) explains it well. 

```
    @inbounds for i=1:4
        bi = bodies[i]
        # For body i, since velocity is the only part of the object
        # changing on each iteration, it's more efficient to update
        # outside the vector.
        iv = bi.v
        for j=i+1:5
            iv = iv .- Δx[k] .* (mag[k] * bodies[j].m)
            bodies[j] = Body(bodies[j].x,
                             bodies[j].v .+ Δx[k] .* (mag[k] * bi.m),
                             bodies[j].m)
            k += 1
        end
        bodies[i] = Body(bi.x, iv, bi.m)
    end
```