@mikowals 2023-12-22
Adding @unroll decorator to the inner loop of `advance` reduces the time to 1.13 seconds.  

Ideally the amount of unrolling could be linked to the alias `NUM_BODIES` but that led to #1542 .  So I have hardcoded `@unroll(4)`.  

I can also rewrite using `unroll` from algorithm for both the outer and inner loops which works fine with the alias but does make the code a bit less readable.